requirements.txt
aiogram
python-dotenv
aiosqlite
sqlalchemy